+++ src/main/kotlin/de/itemis/mps/gradle/downloadJBR/Tasks.kt
@@ -3,13 +3,23 @@ package de.itemis.mps.gradle.downloadJBR
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.*
+import org.gradle.jvm.toolchain.JavaLauncher
+import org.gradle.jvm.toolchain.JavaToolchainService
+import org.gradle.jvm.toolchain.JavaToolchainSpec
+import org.gradle.jvm.toolchain.internal.SpecificInstallationToolchainSpec
 import java.io.File
+import javax.inject.Inject
 
-open class DownloadJbrForPlatform : DefaultTask() {
+open class DownloadJbrForPlatform @Inject constructor(
+    private val objects: ObjectFactory,
+    private val javaToolchainService: JavaToolchainService
+) : DefaultTask() {
 
     @get:Internal
-    internal val jbrDirProperty: DirectoryProperty = project.objects.directoryProperty()
+    internal val jbrDirProperty: DirectoryProperty = objects.directoryProperty()
 
     @get:Internal
     var jbrDir : File
@@ -19,7 +29,7 @@ open class DownloadJbrForPlatform : DefaultTask() {
         }
 
     @get:Internal
-    internal val javaExecutableProperty: RegularFileProperty = project.objects.fileProperty()
+    internal val javaExecutableProperty: RegularFileProperty = objects.fileProperty()
 
     @get:Internal
     var javaExecutable: File
@@ -27,4 +37,18 @@ open class DownloadJbrForPlatform : DefaultTask() {
         set(value) {
             javaExecutableProperty.set(value)
         }
+
+    /**
+     * A [JavaToolchainSpec] that can be passed to [JavaToolchainService] to obtain various tools (Java compiler,
+     * launcher, javadoc).
+     */
+    @get:Internal
+    val toolchainSpec: Provider<JavaToolchainSpec> =
+        javaExecutableProperty.map { SpecificInstallationToolchainSpec.fromJavaExecutable(objects, it.asFile.path) }
+
+    /**
+     * A [JavaLauncher] for the downloaded JBR that can be used with [JavaExec] task.
+     */
+    @get:Internal
+    val javaLauncher: Provider<JavaLauncher> = toolchainSpec.flatMap(javaToolchainService::launcherFor)
 }