+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.27.0
+
+### Added
+
+- `DownloadJbrForPlatform#toolchainSpec`: a `Provider<JavaToolchainSpec>` that can be used with `JavaToolchainService`
+  to look up tools such as the Java launcher, compiler or Javadoc generator.
+- `DownloadJbrForPlatform#javaLauncher`: a `Provider<JavaLauncher>` that can be used with `JavaExec` tasks.
+
 ## 1.26.0
 
 ### Changed